move logging to pkg
