+++ pkg/logging/elastic.go
