+++ pkg/logging/kinesis.go
