+++ pkg/logging/logstash.go
