+++ pkg/logging/s3.go
