+++ pkg/logging/stdout.go
